+++ benchmark/src/libfork_benchmark/fib/fib.hpp
@@ -9,7 +9,8 @@
 
 #include "libfork_benchmark/common.hpp"
 
-import libfork;
+import libfork.core;
+import libfork.schedule;
 
 inline constexpr int fib_test = 3;
 inline constexpr int fib_base = 37;
