--- src/batteries/batteries.cxx
@@ -1,6 +1,0 @@
-export module libfork.batteries;
-
-export import :mono_context;
-export import :adaptors;
-export import :geometric_stack;
-export import :dummy_stack;
