+++ src/core/core.cxx
@@ -6,12 +6,19 @@ export module libfork.core;
 // schedule
 // concepts
 // polymorphic context
+// conformant stacks
 
 export import :concepts;
 export import :constants;
+export import :utility;
+export import :tuple;
+export import :geometric_stack;
+export import :dummy_stack;
 export import :frame;
+export import :deque;
 export import :thread_locals;
 export import :poly_context;
+export import :mono_context;
 export import :ops;
 export import :handles;
 export import :promise;
@@ -19,5 +26,4 @@ export import :schedule;
 export import :root;
 export import :execute;
 export import :receiver;
-
-import libfork.utils;
+export import :adaptors;
