+++ src/core/stacks/adaptor.cxx
