+++ src/core/thread_locals.cxx
@@ -3,7 +3,7 @@ export module libfork.core:thread_locals;
 import std;
 
 import :concepts;
-import libfork.utils;
+import :utility;
 
 namespace lf {
