--- src/libfork.cxx
@@ -1,5 +1,0 @@
-export module libfork;
-
-export import libfork.core;
-export import libfork.batteries;
-export import libfork.schedulers;
