+++ src/schedule/schedule.cxx
@@ -1,0 +1,4 @@
+export module libfork.schedule;
+
+export import :dummy_context;
+export import :inline_scheduler;
