--- src/schedulers/schedulers.cxx
@@ -1,4 +1,0 @@
-export module libfork.schedulers;
-
-export import :dummy;
-export import :inline_scheduler;
