--- src/utils/utils.cxx
@@ -1,6 +1,0 @@
-export module libfork.utils;
-
-export import :utility;
-export import :tuple;
-export import :deque;
-export import :concepts;
