+++ test/src/stack.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 import std;
-import libfork;
+import libfork.core;
 
 using lf::k_new_align;
 using lf::worker_stack;
